requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.4
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
django-resized==1.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
phonenumbers==8.13.18
Pillow==10.0.0
PyJWT==2.8.0
pyotp==2.9.0
pytz==2023.3
sqlparse==0.4.4
typing-extensions==4.7.1
